refactor(edge): move scripts form to edge module
