+++ app/edge/components/EdgeScriptForm/EdgePropertiesForm.tsx
