+++ app/edge/components/EdgeScriptForm/index.ts
@@ -1,0 +1,1 @@
+export { EdgeScriptForm, EdgeScriptFormAngular } from './EdgeScriptForm';
